Revert "Document changes to my refile workflow"
This reverts commit ec1152a880a8d580d8e1c0668fa1457ab8429e47.

This was a bad experiment.  Normally I let stuff cook for 30 days
and that's long enough to know if it works or not but I found
that I was fumbling around too much trying to remember exactly what
some task was called and refiling in steps is more logical for my
setup.... so *poof* there goes the failed refile directly experiment.