Clock in Remember Tasks
All remember tasks now clock in when they are created.
This allows us to use it for timing phone calls and for keeping track
of how much time is spent during the interruption.

To permanently switch to the remember task we just save it and
use f9-SPC to switch the clock to it.