Replace USB stick with dedicated server and new script
I had problems with my USB stick setup where occassionally the stick got
corrupted.  Rebuilding 35+ git repos on the stick after the third time was
getting old.

This replaces the USB stick with a server setup instead.  The new git-sync
script is more user-friendly.

Synchronization goes something like this:

work on the workstation:
   hack hack hack
   commit commit commit
   <time to move to the laptop>
   git-sync
        - updates all repos to the bare repositories
   <do any required manual merges and repeat git-sync until done>

on laptop:
   git-sync
   resolve any issues with merges if any
   <do any required manual merges and repeat git-sync until done>
   walk away with the laptop.

So just git-sync on both ends and then go to work.