Always select default clocking task on punch-in
I normally punch in and only when I'm changing contexts.  This means
I'm switching from working for client A to client B or to some other
non-client work and I want the default clocking task to change to the
next client.  This is more convenient for me than running F12-c and
clocking in the task.  I have f9-I burned into my fingers by now.