Update phone call handling
This is much better.  Now we just use capture mode with a few
functions to populate the details - and it's just like any other
capture template.

Drop the TODO keyword - the call is done normally when I C-c C-c on
the entry (and the clock stops).  If I need to navigate away from the
capture buffer I'll C-c C-c and F9-SPC to restart the clock for the call.