Turn off org-insert-heading-respect-content
The control versions of the heading insert commands temporarily turn
this on.  Sometimes I want to insert stuff in the middle of an entry
and that's difficult when the setting is forced on always.

M-RET and M-S-RET inserts without respecting content
C-RET and C-S-RET inserts respecting content.

I like this better.