org-agenda-files is now a list of directories
Replace the list of files with a list of directories so new files
are automatically picked up without requiring any setup.  This also
allows multiple git repositories of org-files to contribute to the
agenda.

Some machines do not have all of the org files available so we just
drop the unavailable directories in this case.

This makes agenda creation a little slower but it is still very
usable and worth it for the extra flexibility.