Automatically set the ARCHIVE property in parent tasks
When we find tasks to archive - set the ARCHIVE property if it is
missing in the parent LEVEL 1 task.  Since all of our archiveable
tasks are LEVEL 2 we can just assume one level up is the correct one.

If an existing ARCHIVE property exists we will just use it - even if
it does not match the task.

Fix a problem with calculating last month - February only has 28
days... so on March 1st, 31 days ago was sometime in January and we
incorrectly select tasks from February to archive which was just a few
days ago.