Change agenda and clocking behaviour
Punch in with a prefix makes the current task the default clocking task
I almost always want the Organization task so I can punch in without
thinking about where the point it.  After work if I want to punch in I
can specify some other task so I'm not applying time to the work org files.

Clocking out of a task clocks in a higher level parent task in the same
subtree as long as it has a todo keyword.  Otherwise clock in the
default task.  This keeps clock time in the project subtree while
spending time on the project.

Agenda restriction locking is now manual.  Enable the restriction setting
from the agenda with a new keybinding for C-c C-x <.  This restricts
to the current selected task subtree or file if given a C-u prefix.

The block agenda relevant tasks is simplified since it no longer depends on
automatic restriction locking.

Set up a variable for the default clocking task