Agenda restriction lock changes
Set restriction lock from agenda to task directly

This is less surprising - it's easy to move the restriction lock up a
level by visiting the task in the org file and going up and resetting
the lock - in case you want to see the entire project.

Selecting the entire project sometimes has too many tasks in it and I
want to limit the view to part of the subtree.