NEXT list entries are for project tasks only
The Tasks list now shows single NEXT tasks (set automatically by
clocking in) instead of moving these to the NEXT list. The NEXT list
is for moving projects forward.

When not narrowed to a project the Tasks lists shows single tasks, and
when narrowed to a project the Tasks lists shows project tasks.