Try allowing user provided variables to override defaults
This allows uses that tangle the document to provide their defaults
for the load path and org-agenda-files.  This way they can set the
override variables first, and then load the tangled org-mode.el file
to get the published settings in their emacs environment.

The tangled document and the published documents now differ for the
following settings:

  - org-mode load-path
  - contrib load-path
  - org-agenda-files

Thanks for Jeff Stern for this idea and the elisp.