Remove duplicate entries from block agenda
Appointments with todo keywords and fixed times will show
up on time in the daily agenda -- so remove tasks with
dates from the other todo lists in the block agenda.

These are appointments with a date -- like
* TODO Dentist Appointment
  <2012-04-08 Sun>

This removes duplication in the lists since it will show up on time in
the daily agenda and doesn't need to be listed again somewhere else.

Remove HOLD projects from the project lists.  Stuck projects show up
in the stuck project list only, and HOLD projects show up in the
waiting list so we don't need to list it twice.