Make restriction handling more friendly
Allow restrictions for Project, Up one level to work anywhere in the
agenda.  If we're not on a task which should be used for the agenda
then operate on the existing restriction.

If we're not in a valid place to change the restriction then use the
current clocking task as the point to act on.  This allows changes in
the agenda when the point is not on any task (ie first or last line)
and we can still select the project or parent task of the current
clocking task.

Fix file restrictions so that F is ignored if it can't set a file
restriction.