Allow duplicate projects when scheduled
This allows me to quickly see all of my projects at a glance.
Projects are listed either under the "Stuck Projects" agenda or the
"Projects" agenda even if they are scheduled.

The V command to view projects now skips over all of the daily agenda
entries and starts at "Tasks to Refile" so we don't view projects more
than once.